Relatively big changes.
The Scraper now handles most of its own exceptions. Only exceptions that are thrown are only IOExceptions right now as far as I know. Also added a function where the GUI can load the courses into the ListView also made it run in a different thread.  The Scraper will now cache the classes onto the disk.