Course + Calendar update
Course: LocalDate objects are used instead of Date object because it's a better idea.

UI: Implemented the code that takes a list of courses and puts it into the calendar. There's a small bug that I know what's casuing it, but not sure on how to fix it just yet.