Fixed duplication bug.
Fixed the bug where it would duplicate events and also made it so that the code is more efficient. If a course only has class once a day at the same time every day it will add it as one event, otherwise if it had multiple times per day or different times on different days it will add events for each day.