Changed MultiMap back + small bug fix
I had changed the MultiMap to, internally, but a map fo Strings to Sets of values. This change broke some other things that relied on this and if I plan on actually doing this change then i need to change a few things. This change is being considered. Also removing a course didn't subtract it from the credit load, that is not fixed.