Big Changes!
A lot of the changes were mostly just documentation changes. I added lots of documentation for the various things that I've created as well as fixed previous documentation. I also run some code formatting as well. Lastly I added a browser that will display course descriptions. This happens if a user right clicks on a course.